+++ crates/miden-objects/src/batch/proven_batch.rs
@@ -1,7 +1,7 @@
 use alloc::{collections::BTreeMap, vec::Vec};
 
 use vm_core::utils::{ByteReader, ByteWriter, Deserializable, Serializable};
-use vm_processor::DeserializationError;
+use vm_processor::{DeserializationError, Digest};
 
 use crate::{
     account::AccountId,
@@ -15,6 +15,8 @@ use crate::{
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct ProvenBatch {
     id: BatchId,
+    reference_block_commitment: Digest,
+    reference_block_num: BlockNumber,
     account_updates: BTreeMap<AccountId, BatchAccountUpdate>,
     input_notes: InputNotes<InputNoteCommitment>,
     output_notes_smt: BatchNoteTree,
@@ -27,8 +29,11 @@ impl ProvenBatch {
     // --------------------------------------------------------------------------------------------
 
     /// Creates a new [`ProvenBatch`] from the provided parts.
+    #[allow(clippy::too_many_arguments)]
     pub fn new(
         id: BatchId,
+        reference_block_commitment: Digest,
+        reference_block_num: BlockNumber,
         account_updates: BTreeMap<AccountId, BatchAccountUpdate>,
         input_notes: InputNotes<InputNoteCommitment>,
         output_notes_smt: BatchNoteTree,
@@ -37,6 +42,8 @@ impl ProvenBatch {
     ) -> Self {
         Self {
             id,
+            reference_block_commitment,
+            reference_block_num,
             account_updates,
             input_notes,
             output_notes_smt,
@@ -53,9 +60,24 @@ impl ProvenBatch {
         self.id
     }
 
+    /// Returns the commitment to the reference block of the batch.
+    pub fn reference_block_commitment(&self) -> Digest {
+        self.reference_block_commitment
+    }
+
+    /// Returns the number of the reference block of the batch.
+    pub fn reference_block_num(&self) -> BlockNumber {
+        self.reference_block_num
+    }
+
     /// Returns the block number at which the batch will expire.
     pub fn batch_expiration_block_num(&self) -> BlockNumber {
         self.batch_expiration_block_num
+    }
+
+    /// Returns an iterator over the IDs of all accounts updated in this batch.
+    pub fn updated_accounts(&self) -> impl Iterator<Item = AccountId> + use<'_> {
+        self.account_updates.keys().copied()
     }
 
     /// Returns the map of account IDs mapped to their [`BatchAccountUpdate`]s.
@@ -101,6 +123,8 @@ impl ProvenBatch {
 impl Serializable for ProvenBatch {
     fn write_into<W: ByteWriter>(&self, target: &mut W) {
         self.id.write_into(target);
+        self.reference_block_commitment.write_into(target);
+        self.reference_block_num.write_into(target);
         self.account_updates.write_into(target);
         self.input_notes.write_into(target);
         self.output_notes_smt.write_into(target);
@@ -112,6 +136,8 @@ impl Serializable for ProvenBatch {
 impl Deserializable for ProvenBatch {
     fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
         let id = BatchId::read_from(source)?;
+        let reference_block_commitment = Digest::read_from(source)?;
+        let reference_block_num = BlockNumber::read_from(source)?;
         let account_updates = BTreeMap::read_from(source)?;
         let input_notes = InputNotes::<InputNoteCommitment>::read_from(source)?;
         let output_notes_smt = BatchNoteTree::read_from(source)?;
@@ -120,6 +146,8 @@ impl Deserializable for ProvenBatch {
 
         Ok(Self::new(
             id,
+            reference_block_commitment,
+            reference_block_num,
             account_updates,
             input_notes,
             output_notes_smt,